refactor: reorganize data and recommendations structure
- Move recommendations/data to src/data for shared access
- Move recommendations domain into layout/Recommendations
- Update all import paths to reference new locations
- Maintain uppercase Recommendations convention for consistency

Data is now shared at the src level while recommendations
functionality remains within the layout domain.
