+++ src/shared/components/ui/SortingControls/SortingControls.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import { ArrowUpDown } from 'lucide-react';
-import type { SortOption, UseFiltersReturn } from '../../../../domains/recommendations/types/filtering';
+import type {
+  SortOption,
+  UseFiltersReturn,
+} from '../../../../domains/layout/recommendations/types/filtering';
 import styles from './SortingControls.module.css';
 
 interface SortingControlsProps {
